requirements.txt
streamlit
python-dotenv
PyPDF2
streamlit-extras
langchain
Pillow
openai
